models/transformation/requirements.txt
efficientnet==1.1.1
opencv-python==4.10.0.84
pillow-heif==0.17.0
pystow==0.5.4
tensorflow>=2.8.0,<=2.15.0
python-magic==0.4.15
fastapi==0.112.0
uvicorn==0.30.5
minio==7.2.7
pyyaml
python-multipart